Adding the time fix

+++ src/hashl.h
@@ -3,4 +3,4 @@
 #include "globals.h"
 
 void add_hashl(hashed_link *h);
-hashed_link *find_hashl(uint16_t seq_num);
+hashed_link *find_hashl(long long unsigned int seq_num);
